+++ src/test/cli.build.test.ts
@@ -153,33 +153,39 @@ describe('Dev Containers CLI', function () {
 			const testFolder = `${__dirname}/configs/dockerfile-with-features`;
 			const image1 = 'image-1';
 			const image2 = 'image-2';
+			await shellExec(`docker rmi -f ${image1} ${image2}`);
 			const res = await shellExec(`${cli} build --workspace-folder ${testFolder} --image-name ${image1} --image-name ${image2}`);
 			const response = JSON.parse(res.stdout);
 			assert.equal(response.outcome, 'success');
 			assert.equal(response.imageName[0], image1);
 			assert.equal(response.imageName[1], image2);
+			await shellExec(`docker inspect --type image ${image1} ${image2}`);
 		});
 
 		it('should succeed with multiple --image-name parameters when dockerComposeFile is present', async () => {
 			const testFolder = `${__dirname}/configs/compose-Dockerfile-alpine`;
 			const image1 = 'image-1';
 			const image2 = 'image-2';
+			await shellExec(`docker rmi -f ${image1} ${image2}`);
 			const res = await shellExec(`${cli} build --workspace-folder ${testFolder} --image-name ${image1} --image-name ${image2}`);
 			const response = JSON.parse(res.stdout);
 			assert.equal(response.outcome, 'success');
 			assert.equal(response.imageName[0], image1);
 			assert.equal(response.imageName[1], image2);
+			await shellExec(`docker inspect --type image ${image1} ${image2}`);
 		});
 
 		it('should succeed with multiple --image-name parameters when image is present', async () => {
 			const testFolder = `${__dirname}/configs/image`;
 			const image1 = 'image-1';
 			const image2 = 'image-2';
+			await shellExec(`docker rmi -f ${image1} ${image2}`);
 			const res = await shellExec(`${cli} build --workspace-folder ${testFolder} --image-name ${image1} --image-name ${image2}`);
 			const response = JSON.parse(res.stdout);
 			assert.equal(response.outcome, 'success');
 			assert.equal(response.imageName[0], image1);
 			assert.equal(response.imageName[1], image2);
+			await shellExec(`docker inspect --type image ${image1} ${image2}`);
 		});
 
 		it('should fail with --push true and --output', async () => {
@@ -217,7 +223,7 @@ describe('Dev Containers CLI', function () {
 			try {
 				await shellExec(`docker buildx create --name ${builderName} --driver docker-container --use`);
 
-				const testFolder = `${__dirname}/configs/dockerfile-without-features`;
+				const testFolder = `${__dirname}/configs/dockerfile-with-features`;
 				const outputPath = `${os.tmpdir()}/test-build-cache`;
 				const res = await shellExec(`${cli} build --workspace-folder ${testFolder} --log-level trace --cache-to=type=local,dest=${outputPath}`);
 				console.log(res.stdout);
@@ -234,7 +240,7 @@ describe('Dev Containers CLI', function () {
 			try {
 				await shellExec(`docker buildx create --name ${builderName} --driver docker-container --use`);
 
-				const testFolder = `${__dirname}/configs/image`;
+				const testFolder = `${__dirname}/configs/image-with-features`;
 				const outputPath = `${os.tmpdir()}/test-build-cache-image`;
 				const res = await shellExec(`${cli} build --workspace-folder ${testFolder} --log-level trace --cache-to=type=local,dest=${outputPath}`);
 				console.log(res.stdout);