+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Notable changes.
 
+## March 2024
+
+### [0.58.0]
+- Allow empty value for remote env. (https://github.com/devcontainers/ci/issues/231)
+- Add generate-docs subcommand for templates and features. (https://github.com/devcontainers/cli/pull/759)
+- Only use SELinux label for Linux hosts. (https://github.com/devcontainers/cli/issues/776)
+
+### [0.57.0]
+- Fix crash updating UID/GID when the image's platform is different from the native CPU arch (https://github.com/devcontainers/cli/pull/746)
+- Add tags with build command (https://github.com/devcontainers/ci/issues/271)
+
 ## February 2024
 
 ### [0.56.2]